[ refactor ] 把複製邀請連結按鈕變成component
