+++ src/components/Modal/InviteMember.tsx
@@ -3,14 +3,17 @@ import { InviteMemberCss } from "./style";
 import { useApi } from "@/hooks/useApiHook";
 import { newBoardApi } from "@/api/boards";
 import { NewBoardsProps } from "@/interfaces/boards";
-import { Button, Col, Form, Input, Row } from "antd";
+import { Form, Input, Row } from "antd";
 import { LinkOutlined } from "@ant-design/icons";
+import { OrganizationProps } from "@/interfaces/organization";
+import CopyInviteLinkBtn from "@/components/WorkSpace/CopyInviteLinkBtn";
 
 const InviteMember: React.FC<{
   open: boolean;
   setOpen: Function;
   organizationId: string;
-}> = ({ open, setOpen }) => {
+  userOrganization?: OrganizationProps;
+}> = ({ open, setOpen, userOrganization }) => {
   const [_form] = Form.useForm<NewBoardsProps>();
   const onCancel: () => void = () => {
     setOpen(false);
@@ -45,39 +48,29 @@ const InviteMember: React.FC<{
           <Input style={{ height: "36px" }} />
         </Form.Item>
         <Form.Item name="inviteLink">
-          <Row align={"middle"}>
-            <LinkOutlined
-              style={{
-                padding: "12px",
-                backgroundColor: "var(--grayd5)",
-                borderRadius: "2px",
-              }}
-            />
-            <Col style={{ marginLeft: "16px" }}>
-              <Row align={"middle"} justify={"center"}>
-                <p
-                  style={{
-                    fontWeight: "400",
-                    fontSize: "14px",
-                    lineHeight: "22px",
-                  }}
-                >
-                  使用連結邀請某人加入此工作區
-                </p>
-              </Row>
-
-              <Button
+          <Row align={"middle"} justify={"space-between"}>
+            <Row align={"middle"} style={{ gap: "4px" }}>
+              <LinkOutlined
                 style={{
-                  padding: "0",
-                  margin: "0",
-                  color: "var(--grey9F)",
-                  textDecoration: "underline",
+                  padding: "12px",
+                  backgroundColor: "var(--grayd5)",
+                  borderRadius: "2px",
                 }}
-                type="link"
+              />
+              <p
+                style={{
+                  fontWeight: "400",
+                  fontSize: "14px",
+                  lineHeight: "22px",
+                }}
               >
-                建立連結
-              </Button>
-            </Col>
+                使用連結邀請某人加入此工作區
+              </p>
+            </Row>
+            <CopyInviteLinkBtn
+              userOrganization={userOrganization}
+              onCancel={onCancel}
+            />
           </Row>
         </Form.Item>
       </Form>
